+++ src/core/renderer/dx11/DX11.h
@@ -49,7 +49,7 @@ namespace ninniku {
         void Finalize() override;
         bool Initialize() override;
         bool LoadShader(const std::filesystem::path& path) override;
-        bool LoadShader(const std::string_view& name, const void* pData, const size_t size) override;
+        bool LoadShader(const std::string_view& name, const void* pData, const uint32_t size) override;
         MappedResourceHandle Map(const BufferHandle& bObj) override;
         MappedResourceHandle Map(const TextureHandle& tObj, const uint32_t index) override;
         bool UpdateConstantBuffer(const std::string_view& name, void* data, const uint32_t size) override;
